move fret view to own module
