+++ requirements/base.txt
@@ -44,3 +44,5 @@ django-avatar==4.1.0
 numpy==1.13.0  # pyup: ignore
 parasail==1.1.12
 scikit-bio==0.5.4
+
+channels==2.1.3
